task: extend app-template
