+++ packages/cli/src/lifecycles/scaffold.js
@@ -89,7 +89,7 @@ const writeRoutes = async(compilation) => {
       const appDir = path.join(compilation.context.scratchDir, 'app');
 
       await fs.ensureDir(appDir);
-      await fs.writeFile(path.join(appDir, './app.js'), result);
+      await fs.writeFile(path.join(appDir, './appTemplate.js'), result);
 
       resolve();
     } catch (err) {
@@ -97,6 +97,22 @@ const writeRoutes = async(compilation) => {
     }
   });
 };
+
+const writeBaseAppTemplate = async({ context }) => {
+  return new Promise(async (resolve, reject) => {
+    try{
+      let data = await fs.readFile(path.join(__dirname, '../templates/', 'base-template.js'), 'utf8');
+
+      const appDir = path.join(context.scratchDir, 'app');
+
+      await fs.ensureDir(appDir);
+      await fs.writeFile(path.join(appDir, './app.js'), data);
+      resolve();
+    } catch(err) {
+      reject(err);
+    }
+  })
+}
 
 const setupIndex = async({ context }) => {
   return new Promise(async (resolve, reject) => {
@@ -124,6 +140,9 @@ module.exports = generateScaffolding = async (compilation) => {
 
       console.log('Writing Lit routes...');
       await writeRoutes(compilation);
+
+      console.log('Write Base App Template...');
+      await writeBaseAppTemplate(compilation);
 
       console.log('setup index page and html');
       await setupIndex(compilation);
