+++ packages/cli/src/templates/index.html
@@ -19,7 +19,7 @@
 
   <body>
 
-    <eve-app></eve-app>
+    <base-app></base-app>
     
   </body>
